Split Blitter comparison diagnostics
Move comparison-mode diagnostics out of the active Blitter implementation so debug validation code can evolve separately from the core render path.

Made-with: Cursor
